queue/requirements.txt
functions-framework==3.*
google-analytics-admin==0.15.0
google-analytics-data==0.16.1
google-api-core==2.11.0
google-api-python-client==2.82.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-cloud-pubsub==2.17.0
google-cloud-secret-manager==2.16.1
googleapis-common-protos==1.59.0
psycopg2-binary
typing_extensions==4.5.0
SQLAlchemy